Move email into comms
